int overflow fix to match swarmjs clock

+++ swarm4j-core/src/main/java/citrea/swarm4j/core/clocks/AbstractClock.java
@@ -25,7 +25,7 @@ public abstract class AbstractClock implements Clock {
 
     VersionToken lastIssuedTimestamp;
     int lastSeqSeen;
-    int lastTimeSeen;
+    long lastTimeSeen;
 
     protected AbstractClock(String processId, int timePartLen) {
         this.id = processId;
